requirements.txt
django~=4.2.7

djangorestframework~=3.14.0
markdown~=3.5
django-filter~=23.3
django-environ
celery[redis]~=5.3.4
psycopg2-binary
requests~=2.31.0
django-ptpython
responses

drf-spectacular
environ~=1.0
pip~=23.3.1
pathlib~=1.0.1
attrs~=23.1.0
wheel~=0.41.2
mock~=5.1.0
Pillow~=10.1.0
amqp~=5.1.1
vine~=5.0.0
pytz~=2023.3.post1
PyYAML~=6.0.1
typing_extensions~=4.8.0
kombu~=5.3.2
billiard~=4.1.0
redis~=4.6.0
click~=8.1.7
Pygments~=2.16.1
python-dateutil~=2.8.2
pytest~=7.4.3
asgiref~=3.7.2
sqlparse~=0.4.4
colorama~=0.4.6
numpy~=1.26.1
certifi~=2023.7.22
wcwidth~=0.2.8
six~=1.16.0
idna~=3.4
urllib3~=2.0.6
referencing~=0.30.2
jsonschema~=4.19.1
uritemplate~=4.1.1
django-crispy-forms~=2.1
pyperclip~=1.8.2
inflection~=0.5.1
responses~=0.24.1